+++ src/main/java/com/adit/backend/domain/auth/dto/OAuth2UserInfo.java
@@ -4,10 +4,10 @@ import static com.adit.backend.global.error.GlobalErrorCode.*;
 
 import java.util.Map;
 
+import com.adit.backend.domain.auth.exception.AuthException;
 import com.adit.backend.domain.user.entity.User;
 import com.adit.backend.domain.user.enums.Role;
 import com.adit.backend.domain.user.enums.SocialType;
-import com.adit.backend.global.security.jwt.exception.AuthException;
 
 import lombok.Builder;
